moved requestbody handling to uploadLambdaHandler
